Potential fix for code scanning alert no. 19: Multiplication result converted to larger type
To fix the problem, we need to ensure that the multiplication is performed using the larger integer type to avoid overflow. This can be achieved by casting one of the operands to unsigned long before performing the multiplication. This way, the multiplication will be done using unsigned long arithmetic, preventing overflow.

Specifically, we will cast target->codeidx->SIZ.XTnum to unsigned long before multiplying it by target->codeidx->SIZ.YTnum.

+++ libopenjpeg-2.3.1/src/lib/openjpip/cachemodel_manager.c
@@ -225,7 +225,7 @@ OPJ_BOOL is_allsent(cachemodel_param_t cachemodel)
     TPnum = get_nmax(target->codeidx->tilepart);
 
     if (cachemodel.jppstream) {
-        for (i = 0; i < target->codeidx->SIZ.XTnum * target->codeidx->SIZ.YTnum; i++) {
+        for (i = 0; i < (unsigned long)target->codeidx->SIZ.XTnum * target->codeidx->SIZ.YTnum; i++) {
             if (!cachemodel.th_model[i]) {
                 return OPJ_FALSE;
             }
